% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stand_variables.R
\name{add_stand_variables}
\alias{add_stand_variables}
\title{add_stand_variables}
\usage{
add_stand_variables(df)
}
\arguments{
\item{df}{a data frame with individual tree measurements that include basal
area and the upscale factors. All trees should also be described with plotID
and year variables}
}
\value{
a data frame with added stand variables: total stand basal area and
the number of trees per hectare
}
\description{
This function adds two variables to existing data frame of individual tree
measurements: 1) stand basal area and 2) the number of trees per hectare
}
\examples{
data(data_v1)
data_v1 <- add_stand_variables(df = data_v1)

}
