% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stand_variables_halfPeriod.R
\name{add_stand_variables_halfPeriod}
\alias{add_stand_variables_halfPeriod}
\title{add_stand_variables_halfPeriod}
\usage{
add_stand_variables_halfPeriod(df)
}
\value{
a data frame with added stand variables: total stand basal area and
the number of trees per hectare in the middle of a simulation step
}
\description{
Function for the calculation of stand variables: stand basal area and the
number of trees
}
\keyword{internal}
