% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAI_modul.R
\name{BAI_prediction}
\alias{BAI_prediction}
\title{BAI_prediction}
\usage{
BAI_prediction(
  df_fit,
  df_predict,
  species_n_threshold = 100,
  site_vars,
  include_climate,
  eval_model_BAI = TRUE,
  rf_mtry = NULL,
  k = 10,
  blocked_cv = TRUE
)
}
\description{
BAI model for the MLFS
}
\keyword{internal}
