% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAI_modul_halfPeriod.R
\name{BAI_prediction_halfPeriod}
\alias{BAI_prediction_halfPeriod}
\title{BAI_prediction_halfPeriod}
\usage{
BAI_prediction_halfPeriod(
  df_fit,
  df_predict,
  species_n_threshold = 100,
  site_vars,
  include_climate,
  rf_mtry = NULL
)
}
\description{
BAI model for the MLFS to estimate BAI for half period
}
\keyword{internal}
