% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BAL.R
\name{calculate_BAL}
\alias{calculate_BAL}
\title{calculate_BAL}
\usage{
calculate_BAL(df)
}
\description{
Function for the calculation of competition index BAL (Basal area in larger
trees)
}
\keyword{internal}
