% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BAL_halfPeriod.R
\name{calculate_BAL_halfPeriod}
\alias{calculate_BAL_halfPeriod}
\title{calculate_BAL_halfPeriod}
\usage{
calculate_BAL_halfPeriod(df)
}
\description{
Function for the calculation of competition index BAL (Basal area in larger
trees)
}
\keyword{internal}
