% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_standVars.R
\name{calculate_standVars}
\alias{calculate_standVars}
\title{calculate_standVars}
\usage{
calculate_standVars(df)
}
\description{
Function for the calculation of stand variables: stand basal area and the
number of trees
}
\keyword{internal}
