% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_standVars_halfPeriod.R
\name{calculate_standVars_halfPeriod}
\alias{calculate_standVars_halfPeriod}
\title{calculate_standVars_halfPeriod}
\usage{
calculate_standVars_halfPeriod(df)
}
\description{
Function for the calculation of stand variables: stand basal area and the
number of trees
}
\keyword{internal}
