% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_modul.R
\name{crownHeight_prediction}
\alias{crownHeight_prediction}
\title{crownHeight_prediction}
\usage{
crownHeight_prediction(
  df_fit,
  df_predict,
  site_vars = site_vars,
  species_n_threshold = 100,
  k = 10,
  eval_model_crownHeight = TRUE,
  crownHeight_model = "lm",
  BRNN_neurons = 3,
  blocked_cv = TRUE
)
}
\description{
Models to predict crown height
}
\keyword{internal}
