% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_modul_halfPeriod.R
\name{crownHeight_prediction_halfPeriod}
\alias{crownHeight_prediction_halfPeriod}
\title{crownHeight_prediction_halfPeriod}
\usage{
crownHeight_prediction_halfPeriod(
  df_fit,
  df_predict,
  site_vars = site_vars,
  species_n_threshold = 100,
  crownHeight_model = "lm",
  BRNN_neurons = 3
)
}
\value{
a data frame with imputed crown heights in the middle of a simulation step
}
\description{
Models to predict crown height (half period)
}
\keyword{internal}
