% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_climate.R
\docType{data}
\name{data_climate}
\alias{data_climate}
\title{An example of climate data}
\format{
A data frame with 16695 rows and 5 variables:
\describe{
  \item{plotID}{a unique identifier for plot}
  \item{year}{year}
  \item{month}{month}
  \item{t_avg}{monthly mean temperature}
  \item{p_sum}{monthly precipitation sum}
}
}
\usage{
data_climate
}
\description{
This is simulated monthly climate data, and consists of precipitation sum and mean temperature
}
\keyword{datasets}
