% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_final_cut_weights.R
\docType{data}
\name{data_final_cut_weights}
\alias{data_final_cut_weights}
\title{An example of data_final_cut_weights}
\format{
A data frame with 36 rows and 6 variables:
\describe{
  \item{species}{species name as used in data_NFI}
  \item{step_1}{final cut weight applied in step 1}
  \item{step_2}{final cut weight applied in step 2}
  \item{step_3}{final cut weight applied in step 3}
  \item{step_4}{final cut weight applied in step 4}
  \item{step_5}{final cut weight applied in step 5 and all subsequent steps}
}
}
\usage{
data_final_cut_weights
}
\description{
Each species should have one weight that is multiplied with the probability of being harvested when final_cut is applied
}
\keyword{datasets}
