% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_site.R
\docType{data}
\name{data_site}
\alias{data_site}
\title{An example of site descriptors}
\format{
A data frame with 371 rows and 5 variables:
\describe{
  \item{plotID}{a unique identifier for plot}
  \item{slope}{slope on a plot}
  \item{elevation}{plot elevation}
  \item{northness}{plot northness, 1 is north, 0 is south}
  \item{siteIndex}{a proxy for site index, higher value represents more productive sites}
}
}
\usage{
data_site
}
\description{
This is simulated data describing site descriptors
}
\keyword{datasets}
