% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_v2.R
\docType{data}
\name{data_v2}
\alias{data_v2}
\title{An example of joined national forest inventory and site data that is used
within the MLFS}
\format{
A data frame with 6948 rows and 14 variables:
\describe{
  \item{plotID}{a unique identifier for plot}
  \item{treeID}{a unique identifier for tree}
  \item{year}{year in which plot was visited}
  \item{speciesGroup}{identifier for species group}
  \item{code}{status of a tree: 0 (normal), 1(harvested), 2(dead), 3 (ingrowth)}
  \item{species}{species name}
  \item{height}{tree height in meters}
  \item{crownHeight}{crown height in meters}
  \item{BA}{basal area of individual trees in m2}
  \item{weight}{upscale weight to calculate hectare values}
  \item{p_BA}{basal area of individual trees in m2 from previous simulation step}
  \item{p_weight}{upscale weight to calculate hectare values from previous simulation step}
  \item{p_height}{tree height in meters from previous simulation step}
  \item{p_crownHeight}{crown height in meters from previous simulation step}
}
}
\usage{
data_v2
}
\description{
This is simulated data that reassemble the national forest inventory data.
We use it to show how to run examples for tree and crown height predictions
}
\keyword{datasets}
