% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_factors.R
\docType{data}
\name{form_factors}
\alias{form_factors}
\title{An example table with form factors used to calculate tree volume}
\format{
A data frame with 1199 rows and 3 variables:
\describe{
  \item{plotID}{a unique identifier for plot}
  \item{species}{species name as used in data_NFI}
  \item{form}{for factor used to calculate tree volume}
}
}
\usage{
form_factors
}
\description{
Form factors can be specified per species, plot or per species and plot
}
\keyword{datasets}
