% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_modul.R
\name{height_prediction}
\alias{height_prediction}
\title{height_prediction}
\usage{
height_prediction(
  df_fit,
  df_predict,
  species_n_threshold = 100,
  height_model = "naslund",
  BRNN_neurons = 3,
  height_pred_level = 0,
  eval_model_height = TRUE,
  blocked_cv = TRUE,
  k = 10
)
}
\description{
Height model
}
\keyword{internal}
