% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_modul_halfPeriod.R
\name{height_prediction_halfPeriod}
\alias{height_prediction_halfPeriod}
\title{height_prediction_halfPeriod}
\usage{
height_prediction_halfPeriod(
  df_fit,
  df_predict,
  species_n_threshold = 100,
  height_model = "naslund",
  BRNN_neurons = 3,
  height_pred_level = 0
)
}
\arguments{
\item{df_fit}{data frame with data being used for model fit}
}
\value{
a data frame with imputed tree heights in the middle of a simulation step
}
\description{
Height model
}
\keyword{internal}
