% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingrowth_parameter_list.R
\docType{data}
\name{ingrowth_parameter_list}
\alias{ingrowth_parameter_list}
\title{An example data of ingrowth_parameter_list}
\format{
A list with 2 elements:
\describe{
  \item{3}{deciles of DBH distribution for ingrowth category 3}
  \item{15}{deciles of DBH distribution for ingrowth category 15}
}
}
\usage{
ingrowth_parameter_list
}
\description{
This is a list with two ingrowth levels: 3 (inner circle) and 15 (outer
circle). In each list there are deciles of DBH distributions that are used to
simulate DBH for new trees, separately for each ingrowth category
}
\keyword{datasets}
