% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingrowth_table.R
\docType{data}
\name{ingrowth_table}
\alias{ingrowth_table}
\title{An example data of ingrowth_table}
\format{
A data frame with 2 rows and 4 variables:
\describe{
  \item{code}{ingrowth codes}
  \item{DBH_threshold}{a DBH threshold for particular ingrowth category}
  \item{DBH_max}{maximum DBH for a particular ingrowth category}
  \item{weight}{the upscale weight for particular measurement category}
}
}
\usage{
ingrowth_table
}
\description{
Ingrowth table is used within the ingrowth sub model to correctly simulate
different ingrowth levels and associated upscale weights
}
\keyword{datasets}
