% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_size_data.R
\docType{data}
\name{max_size_data}
\alias{max_size_data}
\title{An example of data with maximum allowed BA that is used in the mortality
sub model}
\format{
A data frame with 36 rows and 2 variables:
\describe{
  \item{species}{species name}
  \item{BA_max}{The maximum allowed basal area (BA) for each individual species}
}
}
\usage{
max_size_data
}
\description{
This is simulated max_size_data and used for examples in mortality sub model
}
\keyword{datasets}
