% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_thresholds.R
\docType{data}
\name{measurement_thresholds}
\alias{measurement_thresholds}
\title{An example of measurement_thresholds table}
\format{
A data frame with 2 rows and 2 variables:
\describe{
  \item{DBH_threshold}{a DBH threshold for particular measurement category}
  \item{weight}{the upscale weight for particular measurement category}
}
}
\usage{
measurement_thresholds
}
\description{
An example of measurement_thresholds table resulting from concentric plots as used in Slovenian NFI
}
\keyword{datasets}
