% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting_modul.R
\name{simulate_harvesting}
\alias{simulate_harvesting}
\title{simulate_harvesting}
\usage{
simulate_harvesting(
  df,
  harvesting_sum,
  forest_area_ha,
  harvesting_type = "random",
  final_cut_weight = 1,
  thinning_small_weight = 10
)
}
\description{
Harvesting model for the MLFS
}
\keyword{internal}
