% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_functions_halfPeriod.R
\name{V_general_halfPeriod}
\alias{V_general_halfPeriod}
\title{V_general_halfPeriod}
\usage{
V_general_halfPeriod(df, data_volF_param = data_volF_param)
}
\description{
three-parameter volume functions for the MLFS
}
\keyword{internal}
