% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_form_factors.R
\name{vol_form_factors}
\alias{vol_form_factors}
\title{vol_form_factors}
\usage{
vol_form_factors(
  df,
  form_factors = NULL,
  form_factors_level = "species",
  uniform_form_factor = 0.42
)
}
\description{
two-parameter volume functions for the MLFS
}
\keyword{internal}
