% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_form_factors_halfPeriod.R
\name{volume_form_factors_halfPeriod}
\alias{volume_form_factors_halfPeriod}
\title{volume_form_factors_halfPeriod}
\usage{
volume_form_factors_halfPeriod(
  df,
  form_factors = NULL,
  form_factors_level = "species",
  uniform_form_factor = 0.42
)
}
\value{
a data frame with calculated volume for all trees in the middle of
a simulation step
}
\description{
The calculation of individual tree volume using form factors, which can be
defined per species, per plot, or per species and per plot
}
\keyword{internal}
