% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_functions.R
\name{volume_functions}
\alias{volume_functions}
\title{volume_functions}
\usage{
volume_functions(df, df_volumeF_parameters = NULL)
}
\arguments{
\item{df}{data frame with tree heights and basal areas for individual trees}

\item{df_volumeF_parameters}{data frame with equations and parameters for
n-parametric volume functions}
}
\value{
a data frame with calculated volume for all trees
}
\description{
The calculation of individual tree volume using the n-parameter volume
functions for the MLFS
}
\examples{
library(MLFS)
data(data_v3)
data(df_volume_parameters)

data_v3 <- volume_functions(df = data_v3,
  df_volumeF_parameters = df_volume_parameters)

}
