% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_functions_halfPeriod.R
\name{volume_functions_halfPeriod}
\alias{volume_functions_halfPeriod}
\title{volume_functions_halfPeriod}
\usage{
volume_functions_halfPeriod(df, df_volumeF_parameters = NULL)
}
\description{
n-parameter volume functions for the MLFS (half period)
}
\keyword{internal}
