% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumen_functions_slo_2p.R
\name{volume_merchantable}
\alias{volume_merchantable}
\title{volume_merchantable}
\usage{
volume_merchantable(df)
}
\description{
two-parameter volume functions for the MLFS
}
\keyword{internal}
