% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_tariffs.R
\name{volume_tariffs}
\alias{volume_tariffs}
\title{volume_tariffs}
\usage{
volume_tariffs(df, data_tariffs)
}
\arguments{
\item{df}{data frame with tree heights and basal areas for individual trees}

\item{data_tariffs}{data frame with plot- and species-specific parameters for
the calculations of tree volume}
}
\value{
a data frame with calculated volume for all trees
}
\description{
One-parameter volume functions (tariffs) for the MLFS.
}
\examples{
data(data_v3)
data(data_tariffs)
data_v3 <- volume_tariffs(df = data_v3, data_tariffs = data_tariffs)
}
