% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_functions_slo_2p_whole_tree.R
\name{volume_whole_tree}
\alias{volume_whole_tree}
\title{volume_whole_tree}
\usage{
volume_whole_tree(df)
}
\value{
a data frame with calculated volume for all trees
}
\description{
two-parameter volume functions for the MLFS
}
\keyword{internal}
