% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_functions_slo_2p_whole_tree_halfPeriod.R
\name{volume_whole_tree_halfPeriod}
\alias{volume_whole_tree_halfPeriod}
\title{volume_whole_tree_halfPeriod}
\usage{
volume_whole_tree_halfPeriod(df)
}
\value{
a data frame with calculated volume for all trees in the middle of
a simulation step
}
\description{
two-parameter volume functions for the MLFS
}
\keyword{internal}
