% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BAL.R
\name{calculate_BAL}
\alias{calculate_BAL}
\title{calculate_BAL}
\usage{
calculate_BAL(df)
}
\arguments{
\item{df}{a data frame with individual tree measurements that include basal
area and the upscale factors. All trees should also be described with plotID
and year variables}
}
\value{
a data frame with calculated basal area in large trees (BAL)
}
\description{
This function calculates the competition index BAL (Basal Area in Large trees)
and adds it to the table of individual tree measurements that include basal
area and the upscale factors. All trees should also be described with plotID
and year variables
}
\examples{
data(data_v1)
data_v1 <- calculate_BAL(df = data_v1)

}
