% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tariffs.R
\docType{data}
\name{data_tariffs}
\alias{data_tariffs}
\title{An example of table with one-parametric volume functions (adapted uniform French tariffs)}
\format{
A data frame with 1196 rows and 4 variables:
\describe{
  \item{tarifa_class}{tariff class for a particular species on this plot}
  \item{plotID}{plot identifier}
  \item{species}{species name as used in data_NFI}
  \item{v45}{volume of tree with DBH 45 cm}
}
}
\usage{
data_tariffs
}
\description{
The adapted uniform French tariffs are typically used in Slovenia to determine tree volume based on tree DBH
}
\keyword{datasets}
