% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingrowth_modul.R
\name{predict_ingrowth}
\alias{predict_ingrowth}
\title{predict_ingrowth}
\usage{
predict_ingrowth(
  df_fit,
  df_predict,
  site_vars = site_vars,
  include_climate = include_climate,
  eval_model_ingrowth = TRUE,
  k = 10,
  blocked_cv = TRUE,
  ingrowth_model = "glm",
  rf_mtry = NULL,
  ingrowth_table = NULL
)
}
\description{
Ingrowth model
}
\keyword{internal}
