% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_modul.R
\name{predict_mortality}
\alias{predict_mortality}
\title{predict_mortality}
\usage{
predict_mortality(
  df_fit,
  df_predict,
  df_climate,
  mortality_share = NA,
  mortality_share_type = "volume",
  include_climate,
  site_vars,
  select_months_climate = c(1:12),
  mortality_model = "rf",
  nb_laplace = 0,
  k = 10,
  eval_model_mortality = TRUE,
  blocked_cv = TRUE,
  sim_mortality = TRUE,
  sim_step_years = 5,
  rf_mtry = NULL
)
}
\description{
Mortality model
}
\keyword{internal}
