% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting_modul.R
\name{simulate_harvesting}
\alias{simulate_harvesting}
\title{simulate_harvesting}
\usage{
simulate_harvesting(
  df,
  harvesting_sum,
  harvesting_type = "random",
  share_thinning = 0.8,
  final_cut_weight = 1e+07,
  thinning_small_weight = 1e+05,
  harvest_sum_level = 1,
  plot_upscale_type,
  plot_upscale_factor,
  forest_area_ha
)
}
\description{
Harvesting model for the MLFS
}
\keyword{internal}
