% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(df, include_climate, df_climate, select_months_climate)
}
\value{
a data frame with transformed data, which is used in the next steps
of the MLFS
}
\description{
Initial data transformation
}
\keyword{internal}
