% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_tariffs_halfPeriod.R
\name{volume_tariffs_halfPeriod}
\alias{volume_tariffs_halfPeriod}
\title{volume_tariffs_halfPeriod}
\usage{
volume_tariffs_halfPeriod(df, data_tariffs)
}
\value{
a data frame with calculated volume for all trees in the middle of
a simulation step
}
\description{
One-parameter volume functions (tariffs) for the MLFS (half period)
}
\keyword{internal}
